Worker policy created

+++ app/Http/Controllers/WorkerController.php
@@ -49,11 +49,13 @@ class WorkerController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Worker::class);
         return view('workers.create');
     }
 
     public function store(StoreRequest $request)
     {
+        $this->authorize('create', Worker::class);
         $data = $request->validated();
         $data['is_married'] = isset($data['is_married']);
         Worker::create($data);
@@ -63,10 +65,12 @@ class WorkerController extends Controller
 
     public function edit(Worker $worker)
     {
+        $this->authorize('update', $worker);
         return view('workers.edit', compact('worker'));
     }
 
     public function update(Worker $worker, UpdateRequest $request) {
+        $this->authorize('update', $worker);
         $data = $request->validated();
         $data['is_married'] = isset($data['is_married']);
 
@@ -76,6 +80,7 @@ class WorkerController extends Controller
     }
 
     public function destroy(Worker $worker) {
+        $this->authorize('delete', $worker);
         $worker->delete();
         return redirect()->route('workers.index');
     }
